+++ src/composables/colors.ts
@@ -1,11 +1,10 @@
 import { writable, type Writable } from 'svelte/store';
-import { _default } from './defaults.js';
 import { x11Colors } from '$lib/styles/constant/x11.js';
+import type { Colors } from '$lib/types/mytril.js';
+import { configDefault } from '$lib/presets/config.full.js';
 
 const colorFormatRegex = /^(?:[a-zA-Z]+)(?:\d+|A\d+)$/;
-const colors: Writable<{ [key: string]: string | { [key: string]: string } }> = writable(
-	_default.theme.colors
-);
+const colors: Writable<Colors> = writable(configDefault.colors);
 
 export const setColors = (list: { [key: string]: string | { [key: string]: string } }) => {
 	colors.set(list);
@@ -26,7 +25,7 @@ export const getColor = (color?: string) => {
 			if (match) return `var(--${match[1]}-${match[2]})`;
 		}
 
-		return `var(--${color})`;
+		return `var(-myt-color-${color})`;
 	}
 	return undefined;
 };