+++ src/styles/constant/index.ts
@@ -1,1 +1,0 @@
-export { x11Colors } from './x11.js';