+++ src/styles/js/position.ts
@@ -1,9 +1,9 @@
 import { formatBreakpoint } from '$lib/utils/formater.js';
-import { position } from './_constant.js';
+import { configDefault } from '../../presets/config.full.js';
 
 export const positionClass = (screen: string) => {
 	let css = '';
-	for (const element of position) {
+	for (const element of configDefault.assets!.position!) {
 		css += `${formatBreakpoint(screen)}${element} {\n`;
 		css += `position: ${element};\n`;
 		css += `}\n`;