+++ src/utils/dom.ts
@@ -1,3 +1,4 @@
+// legacy
 export function className(...args: Array<string | undefined>) {
 	return args.filter((cls: string | undefined) => !!cls).join(' ');
 }